mobile: extract top level routing into its own hook
